Detach CLI from main project and convert main project into a class library
