+++ PLRPC.CLI/PLRPC.CLI.csproj
@@ -1,6 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
     <PropertyGroup>
+        <OutputType>Exe</OutputType>
         <SelfContained>true</SelfContained>
         <TargetFramework>net7.0</TargetFramework>
         <RuntimeIdentifiers>win-x64;linux-x64</RuntimeIdentifiers>
@@ -11,17 +12,14 @@
     </PropertyGroup>
 
     <PropertyGroup>
-        <RootNamespace>LBPUnion.PLRPC.GUI</RootNamespace>
-        <AssemblyName>LBPUnion.PLRPC.GUI</AssemblyName>
+        <RootNamespace>LBPUnion.PLRPC.CLI</RootNamespace>
+        <AssemblyName>LBPUnion.PLRPC.CLI</AssemblyName>
         <ImplicitUsings>enable</ImplicitUsings>
         <Nullable>disable</Nullable>
         <SelfContained>true</SelfContained>
     </PropertyGroup>
 
     <ItemGroup>
-        <PackageReference Include="Serilog" Version="3.0.1"/>
-        <PackageReference Include="Serilog.Sinks.Console" Version="4.1.0"/>
-        <PackageReference Include="Eto.Forms" Version="2.7.5"/>
         <ProjectReference Include="..\PLRPC\PLRPC.csproj"/>
     </ItemGroup>
