+++ PLRPC.GUI.Windows/PLRPC.GUI.Windows.csproj
@@ -2,6 +2,7 @@
 
     <PropertyGroup>
         <RootNamespace>LBPUnion.PLRPC.GUI</RootNamespace>
+        <AssemblyName>LBPUnion.PLRPC.GUI.Windows</AssemblyName>
         <ImplicitUsings>enable</ImplicitUsings>
         <Nullable>disable</Nullable>
     </PropertyGroup>
