+++ PLRPC.sln
@@ -12,6 +12,10 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "GUI", "GUI", "{112C0D49-EA4
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Core", "Core", "{074150AA-731C-465B-8599-C41B56657C81}"
 EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "CLI", "CLI", "{81EFDCFC-CC1D-4B7E-A921-C64660B9BEF5}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PLRPC.CLI", "PLRPC.CLI\PLRPC.CLI.csproj", "{93BF7C09-2B70-422B-A16B-58850F3F2169}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -34,11 +38,16 @@ Global
 		{5BC9F0D6-8F49-4EC4-BEA0-1C78DF3E0864}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{5BC9F0D6-8F49-4EC4-BEA0-1C78DF3E0864}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{5BC9F0D6-8F49-4EC4-BEA0-1C78DF3E0864}.Release|Any CPU.Build.0 = Release|Any CPU
+		{93BF7C09-2B70-422B-A16B-58850F3F2169}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{93BF7C09-2B70-422B-A16B-58850F3F2169}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{93BF7C09-2B70-422B-A16B-58850F3F2169}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{93BF7C09-2B70-422B-A16B-58850F3F2169}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{20C3CEC9-6328-43A5-AB59-FDCF9D637C40} = {112C0D49-EA43-4443-A1D5-9BCBBEC7A9DB}
 		{5BC9F0D6-8F49-4EC4-BEA0-1C78DF3E0864} = {112C0D49-EA43-4443-A1D5-9BCBBEC7A9DB}
 		{7B4DD2AF-4912-40C2-B6E4-C8FAE3695680} = {112C0D49-EA43-4443-A1D5-9BCBBEC7A9DB}
 		{60270A0A-A34A-40FB-BF97-F070B1325157} = {074150AA-731C-465B-8599-C41B56657C81}
+		{93BF7C09-2B70-422B-A16B-58850F3F2169} = {81EFDCFC-CC1D-4B7E-A921-C64660B9BEF5}
 	EndGlobalSection
 EndGlobal
