+++ PLRPC/PLRPC.csproj
@@ -17,7 +17,6 @@
     </PropertyGroup>
 
     <PropertyGroup>
-        <OutputType>Exe</OutputType>
         <SelfContained>true</SelfContained>
         <TargetFramework>net7.0</TargetFramework>
         <RuntimeIdentifiers>win-x64;linux-x64</RuntimeIdentifiers>
